refactor(bokhulenkov) change to mvc
