+++ TicTacToeApp/Extensions/ColorExtension.swift
@@ -1,0 +1,21 @@
+//
+//  ColorExtension.swift
+//  TicTacToeApp
+//
+//  Created by Alexander Bokhulenkov on 02.10.2024.
+//
+
+import UIKit
+
+extension UIColor {
+    struct CustomColors {
+        static let backgroundWhite = UIColor(red: 1, green: 1, blue: 1, alpha: 1)
+        static let backgroundBlue = UIColor(red: 0.961, green: 0.969, blue: 1, alpha: 1)
+        static let lightBlue = UIColor(red: 0.902, green: 0.914, blue: 0.976, alpha: 1)
+        static let blue = UIColor(red: 0.518, green: 0.502, blue: 0.831, alpha: 1)
+        static let black = UIColor(red: 0.137, green: 0.161, blue: 0.275, alpha: 1)
+        static let purple = UIColor(red: 0.835, green: 0.757, blue: 0.965, alpha: 1)
+        static let pink = UIColor(red: 0.937, green: 0.553, blue: 0.82, alpha: 1)
+        static let gray = UIColor(red: 0.805, green: 0.814, blue: 0.85, alpha: 1)
+    }
+}
