+++ TicTacToeApp/Extensions/ImageExtension.swift
@@ -1,0 +1,25 @@
+//
+//  ImageExtension.swift
+//  TicTacToeApp
+//
+//  Created by Alexander Bokhulenkov on 02.10.2024.
+//
+
+import UIKit
+
+extension UIImage {
+    struct CustomImage {
+        static let burger = UIImage(named: "Burger")
+        static let cross = UIImage(named: "Cross")
+        static let crossWithBg = UIImage(named: "CrossWithBg")
+        static let crossYelow = UIImage(named: "CrossYelow")
+        static let frenchFries = UIImage(named: "FrenchFries")
+        static let heart = UIImage(named: "Heart")
+        static let iceCream = UIImage(named: "IceCream")
+        static let nought = UIImage(named: "Nought")
+        static let noughtGreen = UIImage(named: "NoughtGreen")
+        static let noughtWithBg = UIImage(named: "NoughtWithBg")
+        static let pie = UIImage(named: "Pie")
+        static let star = UIImage(named: "Star")
+    }
+}
